FIXED SHAME
Changed the destory method on the pokemon class to not use the local variable because it was redundant.
Changed to the destroy! method because I'm not validating any parameters. Pushed the nature.name down the stack
to the pokemon model. Renamed it to nature_name instead of just def nature because it was giving me an
infinite loop.