Added EV table, gave a few species ev values, separated pokemon dropdown
Added effort value table to the database. I chose to use stat and then amount because there
will be 5 different stats and for each of those stats the values are 1-4. Then I associated a
few species with evs because for some reason there is a field for that on the api, but there is
no data actually being returned from the api. I only created ev values for those pokemon who I
found from a guide that are recommedned for hoard battles.

For the pokemon dropdown, I am using login in the views at the moment. This should be pushed down
to the model. The species model is scoped, but I'm populating the drop down from the parser service.
I will have to change the parser service to handle this separation. In the pokemon controller I have
to account for the now three different dropdown fields. I used select_tags because when it was using
the form fors, it was tied to the species model. The last field was also overwriting the value with nil
when you chose anything but the last dropdown. Is functional right now but should be refactored.