UX changes to the new and show page. Changed dropdowns on new page.
UX changes on the new page where aligning the dropdowns to the existing columns. On the show page
I added a table to better lay out the stats. On the index page increased the image size and tried
my best to vertically align the elements. For the dropdowns on the new page, instead of using the parsed
data directly from the api, used the Species objects that are already in the database. That way I could
instead use the alphabet scopes defined earlier.