Added Angular for adding evs on pokemon show page.
On the pokemon show page, implemented Angular for the buttons for adding evs. Added
a directive to be able to pass data into it, and the link is then what the directive does.
The template is what gets rendered onto the page, which in this case is the button which we
watch the click event for.